mobx: Move ResultItem and SearchResultSection models out of the SearchStore
