+++ public/app/stores/SearchStore/SearchStore.ts
@@ -1,38 +1,5 @@
 import { types } from 'mobx-state-tree';
-
-export const ResultItem = types.model('ResultItem', {
-  id: types.identifier(types.number),
-  folderId: types.optional(types.number, 0),
-  title: types.string,
-  url: types.string,
-  icon: types.string,
-  folderTitle: types.optional(types.string, ''),
-});
-
-export const SearchResultSection = types
-  .model('SearchResultSection', {
-    id: types.identifier(),
-    title: types.string,
-    icon: types.string,
-    expanded: types.boolean,
-    items: types.array(ResultItem),
-  })
-  .actions(self => ({
-    toggle() {
-      self.expanded = !self.expanded;
-
-      for (let i = 0; i < 100; i++) {
-        self.items.push(
-          ResultItem.create({
-            id: i,
-            title: 'Dashboard ' + self.items.length,
-            icon: 'gicon gicon-dashboard',
-            url: 'asd',
-          })
-        );
-      }
-    },
-  }));
+import { SearchResultSection } from './SearchResultSection';
 
 export const SearchStore = types
   .model('SearchStore', {
